TASK 1 interfaces_execicios - correção

+++ Interfaces_Exercicios/resumo_pedido/Pedido.java
@@ -1,11 +1,11 @@
 import provedores.Frete;
 public class Pedido {
     private int codigo;
-    private int peso;
+    private double peso;
     private double total;
     private Frete frete;
 
-    public Pedido(int codigo, int peso, double total) {
+    public Pedido(int codigo, double peso, double total) {
         this.codigo = codigo;
         this.peso = peso;
         this.total = total;
@@ -19,11 +19,11 @@ public class Pedido {
         this.codigo = codigo;
     }
 
-    public int getPeso() {
+    public double getPeso() {
         return peso;
     }
 
-    public void setPeso(int peso) {
+    public void setPeso(double peso) {
         this.peso = peso;
     }
